# packages/shared-openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lumen API
  version: 0.1.0
servers:
  - url: /api/v1
paths:
  /health/live:
    get:
      tags: [health]
      operationId: getLiveness
      responses:
        "200":
          description: Process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /health/ready:
    get:
      tags: [health]
      operationId: getReadiness
      responses:
        "200":
          description: Process is ready to serve traffic.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadinessResponse"
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Rotating access and refresh token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /auth/refresh:
    post:
      tags: [auth]
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Rotated token pair.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Session revoked.
  /auth/me:
    get:
      tags: [auth]
      operationId: getCurrentPrincipal
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current principal.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrincipalResponse"
  /users:
    get:
      tags: [users]
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users.
    post:
      tags: [users]
      operationId: createUser
      security:
        - bearerAuth: []
      responses:
        "201":
          description: User created.
  /api-keys:
    get:
      tags: [api-keys]
      operationId: listApiKeys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: API keys.
    post:
      tags: [api-keys]
      operationId: createApiKey
      security:
        - bearerAuth: []
      responses:
        "201":
          description: API key created. The secret value is returned once.
  /licenses:
    get:
      tags: [licenses]
      operationId: listLicenses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Licenses.
    post:
      tags: [licenses]
      operationId: createLicense
      security:
        - bearerAuth: []
      responses:
        "201":
          description: License created.
  /nodes:
    get:
      tags: [nodes]
      operationId: listNodes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: VPN nodes.
    post:
      tags: [nodes]
      operationId: createNode
      security:
        - bearerAuth: []
      responses:
        "201":
          description: VPN node registered.
  /subscriptions:
    get:
      tags: [subscriptions]
      operationId: listSubscriptions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subscriptions.
    post:
      tags: [subscriptions]
      operationId: createSubscription
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Subscription created.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ValidationError:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorPayload:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorPayload"
    HealthResponse:
      type: object
      required: [status, checked_at]
      properties:
        status:
          const: ok
        checked_at:
          type: string
          format: date-time
    ReadinessResponse:
      type: object
      required: [status, dependencies]
      properties:
        status:
          enum: [ok, degraded]
        dependencies:
          type: object
          additionalProperties:
            type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          writeOnly: true
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          format: password
          writeOnly: true
    TokenPairResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_at]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          const: Bearer
        expires_at:
          type: string
          format: date-time
    PrincipalResponse:
      type: object
      required: [subject, email, roles, permissions]
      properties:
        subject:
          type: string
          format: uuid
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
    Role:
      enum: [owner, admin, support, node, user]
    Permission:
      enum:
        - api_key:manage
        - license:manage
        - node:manage
        - subscription:read
        - subscription:manage
        - user:manage